Task 4.1 Summary
Created an .env file to hide fine-grained token.
Added token to key and value in Vercel environment.

Created my README File to include the following:
Convert scenarios using this format:
As a [role],
I should be able to [action]
So that [benefit]

After that I converted the following scenarios into Gherkin (Given-When-Then) Syntax. 1-6

Deployed project to Vercel.