4.3 Task - Step 1 Update serverless yml with new function
Update serverless.yml file with new function -- getCalendarEvents

getCalendarEvents:: This is the new key that defines your third Lambda function. It's at the same indentation level as getAuthURL and getAccessToken.
handler: handler.getCalendarEvents: This line tells Serverless Framework that the code for this function is the getCalendarEvents function exported from your handler.js file.
events:: This section defines what triggers the function. In this case, it's an HTTP request.
path: api/get-events/{access_token}: This sets the specific URL endpoint for the function. The {access_token} part in curly braces defines a path parameter. This means the access token will be part of the URL path itself (e.g., .../api/get-events/ya29.a0Af...). Your handler.js code will be able to extract this value from the event object.
method: get: This specifies that the endpoint will only respond to HTTP GET requests, which is appropriate for retrieving a list of events.