Setting up my project with Jest Step 2 out of 5 - Create jest.config.cjs file
In the root of my project I created a new file named jest.config.cjs, Than added the following code into that file.
This file is considered the brain of Jest. It provides the runner with precise instructions on how to locate, transform, and execute the tests within the context of my project.

.cjs stands for "CommonJS Module"