Feat(testing): Add jest-cucumber for BDD setup
Installs the necessary dependencies to begin writing Behavior-Driven
Development (BDD) tests for the Meet App.

This change adds `jest-cucumber` as a dev dependency, which integrates
Gherkin feature files with the Jest testing framework. This setup will
be used for User Acceptance & End-to-End testing, allowing us to
write tests in plain English.

After install received npm warn deprecated --> reflect-metadata@0.2.1 : This version has a critical bug in fallback. Please update to reflect-metadata@0.2.2 or newer.
To do this step run npm update --> do this if error interferes with project.

As of now this does not affect my project. Safe to proceed forward.